refactor: move contents of QueryParameters to Constants and rename DefaultClientOptions
